ml_project/requirements.txt
PyYAML==6.0
pytest==7.1.3
pandas==1.3.5
numpy==1.21.4
scikit-learn==1.0.1